Got float writing to work

+++ src/fake_gpu_thread.c
@@ -92,13 +92,12 @@ static void *run(hashpipe_thread_args_t * args)
         // Write data to shared memory
         int i;
         for (i = 0; i < BIN_SIZE; i++) {
-//             fprintf(stderr, "i: %d\n", i);
-//             fprintf(stderr, "block_idx: %d\n", block_idx);
-//             fprintf(stderr, "NUM_ANTENNAS: %d\n", NUM_ANTENNAS);
-//             fprintf(stderr, "writing: %f\n", (float)(i + (block_idx * BIN_SIZE)));
-            db->block[block_idx].data[i] = (float) (i + (block_idx * BIN_SIZE));
+//             fprintf(stderr, "writing real: %f\n", (float)(i + (block_idx * BIN_SIZE)));
+//             fprintf(stderr, "writing imaginary: %f\n", (float)(i + 1 + (block_idx * BIN_SIZE)));
+//             db->block[block_idx].data[i].real = (float) (i + (block_idx * BIN_SIZE));
+//             db->block[block_idx].data[i].imaginary = (float) (i + 1 + (block_idx * BIN_SIZE));
+            db->block[block_idx].data[i] = i + (block_idx * BIN_SIZE);
 //             db->block[block_idx].data[i] = rand();
-//             db->block[block_idx].data[i] = 0.0f;
         }
 
         // Mark block as full
